clone to new source
