+++ src/en/sflix/build.gradle
@@ -1,13 +1,9 @@
 ext {
     extName = 'SFlix'
     extClass = '.SFlix'
-    extVersionCode = 39
-    isNsfw = false
+    themePkg = 'dopeflix'
+    baseUrl = 'https://sflix.to'
+    overrideVersionCode = 17
 }
 
 apply from: "$rootDir/common.gradle"
-
-dependencies {
-    implementation(project(':lib:playlistutils'))
-    implementation(project(':lib:vidsrcextractor'))
-}
