+++ src/en/sflixfmovie/build.gradle
@@ -1,0 +1,13 @@
+ext {
+    extName = 'SFlix (Unoriginal)'
+    extClass = '.SFlixFmovie'
+    extVersionCode = 1
+    isNsfw = false
+}
+
+apply from: "$rootDir/common.gradle"
+
+dependencies {
+    implementation(project(':lib:playlistutils'))
+    implementation(project(':lib:vidsrcextractor'))
+}
